requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.12.1
twilio==8.10.0
httpx==0.25.1
python-dotenv==1.0.0
pytz==2023.3
email-validator==2.1.0
bcrypt==4.1.0